backend/requirements.txt

# FastAPI e servidor web
fastapi==0.112.0
uvicorn[standard]==0.30.0

# Processamento de arquivos
python-multipart==0.0.6

# Processamento de PDF (múltiplas bibliotecas para fallback)
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105

# OCR para PDFs escaneados
pdf2image==1.16.3
pytesseract>=0.3.13  # Updated for Python 3.14+ compatibility
Pillow>=10.3.0

# Multi-format document extraction
openpyxl>=3.1.0  # Excel .xlsx files
xlrd>=2.0.1  # Excel .xls files (legacy)
python-docx>=1.1.0  # Word .docx files

# AI Integration
openai>=1.101.0
google-genai==1.4.0
google-generativeai>=0.8.0

# Configuração e utilitários
python-dotenv==1.0.0
pydantic>=2.5.0

# LangExtract removido temporariamente (conflitos de dependência)
# langextract==1.0.8

# Dependências opcionais para funcionalidades avançadas
requests==2.31.0
httpx>=0.24,<1.0
psutil==5.9.0

# Database e autenticação
psycopg2-binary>=2.9.10
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Supabase integration
supabase>=2.12.0
